[rocroller] Add Assign nodes in-place of ComputeIndex nodes
