--- shared/rocroller/lib/include/rocRoller/KernelGraph/Transforms/AddComputeIndex.hpp
@@ -1,61 +1,0 @@
-/*******************************************************************************
- *
- * MIT License
- *
- * Copyright 2024-2025 AMD ROCm(TM) Software
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- *******************************************************************************/
-
-#pragma once
-#include <rocRoller/KernelGraph/Transforms/GraphTransform.hpp>
-
-namespace rocRoller
-{
-    namespace KernelGraph
-    {
-        /**
-         * @brief Rewrite KernelGraph to add ComputeIndex operations.
-         *
-         * The ComputeIndex operations are used so that indices do not
-         * need to be completely recalculated everytime when iterating
-         * through a tile of data.
-         *
-         * ComputeIndex operations are added before For Loops and
-         * calculate the first index in the loop.
-         *
-         * A new ForLoopIncrement is added to the loop as well to
-         * increment the index by the stride amount.
-         *
-         * Offset, Stride and Buffer edges are added to the DataFlow
-         * portion of the Coordinate graph to keep track of the data
-         * needed to perform the operations.
-         */
-        class AddComputeIndex : public GraphTransform
-        {
-        public:
-            KernelGraph apply(KernelGraph const& original) override;
-            std::string name() const override
-            {
-                return "AddComputeIndex";
-            }
-        };
-    }
-}
