+++ shared/rocroller/lib/source/KernelGraph/ControlToCoordinateMapper.cpp
@@ -239,41 +239,12 @@ namespace rocRoller::KernelGraph
             return std::visit(rocRoller::overloaded{
                                   [](std::monostate const&) { return "none"; },
                                   [](JustNaryArgument const&) { return "NaryArgument"; },
-                                  [](ComputeIndex const&) { return "ComputeIndex"; },
                                   [](TypeAndSubDimension const&) { return "TypeAndSubDimension"; },
                                   [](TypeAndNaryArgument const&) { return "TypeAndNaryArgument"; },
                                   [](UnrollStride const&) { return "UnrollStride"; },
                                   [](BaseOffset const&) { return "BaseOffset"; },
                               },
                               cs);
-        }
-
-        std::string toString(ComputeIndexArgument cia)
-        {
-            switch(cia)
-            {
-            case ComputeIndexArgument::TARGET:
-                return "TARGET";
-            case ComputeIndexArgument::INCREMENT:
-                return "INCREMENT";
-            case ComputeIndexArgument::BASE:
-                return "BASE";
-            case ComputeIndexArgument::OFFSET:
-                return "OFFSET";
-            case ComputeIndexArgument::STRIDE:
-                return "STRIDE";
-            case ComputeIndexArgument::BUFFER:
-                return "BUFFER";
-            case ComputeIndexArgument::BASEADDRESS:
-                return "BASEADDRESS";
-            default:
-                return "Invalid";
-            }
-        }
-
-        std::ostream& operator<<(std::ostream& stream, ComputeIndexArgument const& cia)
-        {
-            return stream << toString(cia);
         }
 
         struct CSToStringVisitor
@@ -286,11 +257,6 @@ namespace rocRoller::KernelGraph
             std::string operator()(JustNaryArgument const& n) const
             {
                 return toString(n.argument);
-            }
-
-            std::string operator()(ComputeIndex const& ci) const
-            {
-                return concatenate(ci.argument, ": (", ci.index, ")");
             }
 
             std::string operator()(TypeAndSubDimension const& ci) const
