+++ shared/rocroller/lib/source/KernelGraph/Reindexer.cpp
@@ -182,7 +182,6 @@ namespace rocRoller
             }
 
             MAKE_OPERATION_VISITOR(Barrier);
-            MAKE_OPERATION_VISITOR(ComputeIndex);
             MAKE_OPERATION_VISITOR(Deallocate);
             MAKE_OPERATION_VISITOR(DoWhileOp);
             MAKE_OPERATION_VISITOR(Exchange);
