+++ shared/rocroller/lib/source/KernelGraph/Transformations/AddDirect2LDS.cpp
@@ -65,12 +65,11 @@ namespace rocRoller
                     }
                     else
                     {
-                        AssertFatal(
-                            storeLDSTags.size() >= 2,
-                            "AddDirect2LDS: At least 2 ComputeIndex operations required for "
-                            "StoreLDSTile.",
-                            ShowValue(loadGlobal),
-                            ShowValue(storeLDSTags.size()));
+                        AssertFatal(storeLDSTags.size() >= 2,
+                                    "AddDirect2LDS: At least 2 Assign operations required for "
+                                    "StoreLDSTile.",
+                                    ShowValue(loadGlobal),
+                                    ShowValue(storeLDSTags.size()));
                         for(const auto& storeLDS : storeLDSTags)
                         {
                             auto maybeForLoopOfLoad
