+++ shared/rocroller/test/catch/AddDeallocateTest.cpp
@@ -82,7 +82,8 @@ namespace AddDeallocateTest
         transforms.push_back(std::make_shared<OrderEpilogueBlocks>());
         transforms.push_back(std::make_shared<CleanLoops>());
         transforms.push_back(std::make_shared<AddPrefetch>(params, context.get()));
-        transforms.push_back(std::make_shared<AddComputeIndex>());
+        transforms.push_back(
+            std::make_shared<AssignIndexExpressions>(context.get(), example.getCommand()));
 
         for(auto& t : transforms)
             kgraph = kgraph.transform(t);
