+++ shared/rocroller/test/catch/LoadPackedTest.cpp
@@ -56,7 +56,7 @@ TEST_CASE("LoadPacked", "[kernel-graph]")
         std::make_shared<AddLDS>(params, context.get()),
         std::make_shared<LowerLinear>(context.get()),
         std::make_shared<LowerTile>(params, context.get()),
-        std::make_shared<AddComputeIndex>(),
+        std::make_shared<AssignIndexExpressions>(context.get(), example.getCommand()),
         std::make_shared<UpdateWavefrontParameters>(params),
     };
