+++ shared/rocroller/test/unit/LDSTileCopyTest.cpp
@@ -59,45 +59,45 @@ namespace LDSCopyTest
      *
      * Kernel
      *   |
-     *   --[body]--> Scope --[body]--> ComputeIndex --[seq]--> ComputeIndex --[seq]--> LoadTiled
+     *   --[body]--> Scope --[body]--> Assign --[seq]--> Assign --[seq]--> LoadTiled
      *                 |
      *                 |[seq]
      *                 |
      *                 v
-     *              Barrier --[seq]--> Scope --[seq]--> ComputeIndex --[seq]--> ComputeIndex --[seq] -->StoreLDSTiled
+     *              Barrier --[seq]--> Scope --[seq]--> Assign --[seq]--> Assign --[seq] -->StoreLDSTiled
      *
      *
      * New control graph:
      *
      * Kernel
      *   |
-     *   --[body]--> Scope --[body]--> ComputeIndex --[seq]--> ComputeIndex --[seq]--> NOP
-     *                                                                                  |
-     *                                                                                  |[seq]
-     *                                                                                  |
-     *                                                                                  v
-     *                                                                             ComputeIndex
-     *                                                                                  |
-     *                                                                                  |[seq]
-     *                                                                                  |
-     *                                                                                  v
-     *                                                                             ComputeIndex
-     *                                                                                  |
-     *                                                                                  |[seq]
-     *                                                                                  |
-     *                                                                                  v
-     *                                                                                 NOP
-     *                                                                                  |
-     *                                                                                  |[seq]
-     *                                                                                  |
-     *                                                                                  v
-     *                                                                                Barrier
-     *                                                                                  |
-     *                                                                                  |[seq]
-     *                                                                                  |
-     *                                                                                  v
-     *                                                                          LoadTileDirect2LDS
-    */
+     *   --[body]--> Scope --[body]--> Assign --[seq]--> Assign --[seq]--> NOP
+     *                                                                       |
+     *                                                                       |[seq]
+     *                                                                       |
+     *                                                                       v
+     *                                                                     Assign
+     *                                                                       |
+     *                                                                       |[seq]
+     *                                                                       |
+     *                                                                       v
+     *                                                                     Assign
+     *                                                                       |
+     *                                                                       |[seq]
+     *                                                                       |
+     *                                                                       v
+     *                                                                      NOP
+     *                                                                       |
+     *                                                                       |[seq]
+     *                                                                       |
+     *                                                                       v
+     *                                                                     Barrier
+     *                                                                       |
+     *                                                                       |[seq]
+     *                                                                       |
+     *                                                                       v
+     *                                                               LoadTileDirect2LDS
+     */
     void addDirect2LDS(rocRoller::KernelGraph::KernelGraph& kgraph)
     {
         auto loadTiledNodes    = kgraph.control.getNodes<LoadTiled>().to<std::vector>();
@@ -270,12 +270,6 @@ namespace LDSCopyTest
         auto updateWavefrontParams = std::make_shared<UpdateWavefrontParameters>(params);
         kgraph                     = kgraph.transform(updateWavefrontParams);
 
-        auto addComputeIndex = std::make_shared<AddComputeIndex>();
-        kgraph               = kgraph.transform(addComputeIndex);
-
-        if(m_context->kernelOptions()->removeSetCoordinate)
-            kgraph = kgraph.transform(std::make_shared<RemoveSetCoordinate>());
-
         auto command = std::make_shared<rocRoller::Command>();
         command->allocateArgument({DataType::UInt32, PointerType::PointerGlobal},
                                   Operations::OperationTag(0),
@@ -287,8 +281,12 @@ namespace LDSCopyTest
                                   ArgumentType::Value,
                                   DataDirection::ReadOnly,
                                   "a");
+        auto assignIndexExprs = std::make_shared<AssignIndexExpressions>(m_context, command);
+        kgraph                = kgraph.transform(assignIndexExprs);
 
-        kgraph = kgraph.transform(std::make_shared<AssignComputeIndex>(m_context, command));
+        if(m_context->kernelOptions()->removeSetCoordinate)
+            kgraph = kgraph.transform(std::make_shared<RemoveSetCoordinate>());
+
         kgraph = kgraph.transform(std::make_shared<CleanArguments>(m_context, command));
 
         m_context->schedule(k->preamble());
